// src/generated/highcharts-options.gen.ts
export const boostOptionsSchema = z.object({
    allowForce: z.boolean().optional(),
    debug: boostDebugOptionsSchema.optional(),
    enabled: z.boolean().optional(),
    pixelRatio: z.number().optional(),
    seriesThreshold: z.number().optional(),
    useGPUTranslations: z.boolean().optional(),
    usePreallocated: z.boolean().optional()
});

export const captionOptionsSchema = z.object({
    align: z.any().optional(),
    floating: z.boolean().optional(),
    margin: z.number().optional(),
    style: z.any().optional(),
    text: z.string().optional(),
    useHTML: z.boolean().optional(),
    verticalAlign: z.any().optional(),
    x: z.number().optional(),
    y: z.number().optional()
});
// ...
export const creditsOptionsSchema = z.object({
    enabled: z.boolean().optional(),
    href: z.string().optional(),
    mapText: z.string().optional(),
    mapTextFull: z.string().optional(),
    position: z.any().optional(),
    style: z.any().optional(),
    text: z.string().optional()
});
// ...
export const dataOptionsSchema = z.object({
    beforeParse: z.any().optional(),
    columns: z.any().optional()
});
// ...
export const drilldownOptionsSchema = z.object({
    activeAxisLabelStyle: z.any().optional(),
    activeDataLabelStyle: z.any().optional(),
    allowPointDrilldown: z.boolean().optional(),
    animation: z.any().optional(),
    breadcrumbs: drilldownBreadcrumbsOptionsSchema.optional(),
    drillUpButton: drilldownDrillUpButtonOptionsSchema.optional(),
    mapZooming: z.boolean().optional(),
    series: z.any().optional()
});
// ...
export const globalOptionsSchema = z.object({
    buttonTheme: globalButtonThemeOptionsSchema.optional()
});
// ...
export const loadingOptionsSchema = z.object({
    hideDuration: z.number().optional(),
    labelStyle: z.any().optional(),
    showDuration: z.number().optional(),
    style: z.any().optional()
});
// ...
export const mapNavigationOptionsSchema = z.object({
    buttonOptions: mapNavigationButtonOptionsSchema.optional(),
    buttons: mapNavigationButtonsOptionsSchema.optional(),
    enableButtons: z.boolean().optional(),
    enabled: z.boolean().optional(),
    enableDoubleClickZoom: z.boolean().optional(),
    enableDoubleClickZoomTo: z.boolean().optional(),
    enableMouseWheelZoom: z.boolean().optional(),
    enableTouchZoom: z.boolean().optional(),
    mouseWheelSensitivity: z.number().optional()
});
// ...
export const mapViewOptionsSchema = z.object({
    center: z.any().optional(),
    fitToGeometry: z.record(z.string(), z.any()).optional(),
    insetOptions: mapViewInsetOptionsSchema.optional(),
    insets: z.any().optional(),
    maxZoom: z.any().optional(),
    padding: z.any().optional(),
    projection: z.any().optional(),
    zoom: z.number().optional()
});
// ...
export const accessibilityOptionsSchema = z.object({
    announceNewData: z.any().optional(),
    customComponents: z.any().optional(),
    description: z.string().optional(),
    enabled: z.boolean().optional(),
    highContrastMode: z.string().optional(),
    highContrastTheme: z.any().optional(),
    keyboardNavigation: z.any().optional(),
    landmarkVerbosity: z.any().optional(),
    linkedDescription: z.any().optional(),
    point: accessibilityPointOptionsSchema.optional(),
    screenReaderSection: accessibilityScreenReaderSectionOptionsSchema.optional(),
    series: accessibilitySeriesOptionsSchema.optional(),
    typeDescription: z.string().optional()
});

export const chartOptionsSchema = z.object({
    alignThresholds: z.boolean().optional(),
    alignTicks: z.boolean().optional(),
    allowMutatingData: z.boolean().optional(),
    animation: z.any().optional(),
    axisLayoutRuns: z.number().optional(),
    backgroundColor: z.any().optional(),
    borderColor: z.any().optional(),
    borderRadius: z.number().optional(),
    borderWidth: z.number().optional(),
    className: z.string().optional(),
    colorCount: z.number().optional(),
    displayErrors: z.boolean().optional(),
    events: chartEventsOptionsSchema.optional(),
    height: z.any().optional(),
    ignoreHiddenSeries: z.boolean().optional(),
    inverted: z.boolean().optional(),
    map: z.any().optional(),
    mapTransforms: z.any().optional(),
    margin: z.any().optional(),
    marginBottom: z.number().optional(),
    marginLeft: z.number().optional(),
    marginRight: z.number().optional(),
    marginTop: z.number().optional(),
    numberFormatter: z.any().optional(),
    options3d: chart3dOptionsSchema.optional(),
    panKey: z.any().optional(),
    panning: chartPanningOptionsSchema.optional(),
    parallelAxes: z.any().optional(),
    parallelCoordinates: z.boolean().optional(),
    plotBackgroundColor: z.any().optional(),
    plotBackgroundImage: z.string().optional(),
    plotBorderColor: z.any().optional(),
    plotBorderWidth: z.number().optional(),
    plotShadow: z.any().optional(),
    polar: z.boolean().optional(),
    proj4: z.any().optional(),
    reflow: z.boolean().optional(),
    renderTo: z.any().optional(),
    resetZoomButton: chartResetZoomButtonOptionsSchema.optional(),
    scrollablePlotArea: chartScrollablePlotAreaOptionsSchema.optional(),
    selectionMarkerFill: z.any().optional(),
    seriesGroupShadow: z.any().optional(),
    shadow: z.any().optional(),
    showAxes: z.boolean().optional(),
    spacing: z.any().optional(),
    spacingBottom: z.number().optional(),
    spacingLeft: z.number().optional(),
    spacingRight: z.number().optional(),
    spacingTop: z.number().optional(),
    style: z.any().optional(),
    styledMode: z.boolean().optional(),
    type: z.string().optional(),
    width: z.any().optional(),
    zooming: chartZoomingOptionsSchema.optional()
});

export const connectorsOptionsSchema = z.object({
    algorithmMargin: z.number().optional(),
    dashStyle: z.any().optional(),
    enabled: z.boolean().optional(),
    endMarker: connectorsEndMarkerOptionsSchema.optional(),
    lineColor: z.string().optional(),
    lineWidth: z.number().optional(),
    marker: connectorsMarkerOptionsSchema.optional(),
    radius: z.number().optional(),
    startMarker: connectorsStartMarkerOptionsSchema.optional(),
    type: z.any().optional()
});
// ...
export const legendOptionsSchema = z.object({
    accessibility: legendAccessibilityOptionsSchema.optional(),
    align: z.any().optional(),
    alignColumns: z.boolean().optional(),
    backgroundColor: z.any().optional(),
    borderColor: z.any().optional(),
    borderRadius: z.number().optional(),
    borderWidth: z.number().optional(),
    bubbleLegend: legendBubbleLegendOptionsSchema.optional(),
    className: z.string().optional(),
    enabled: z.boolean().optional(),
    events: legendEventsOptionsSchema.optional(),
    floating: z.boolean().optional(),
    itemCheckboxStyle: z.any().optional(),
    itemDistance: z.number().optional(),
    itemHiddenStyle: z.any().optional(),
    itemHoverStyle: z.any().optional(),
    itemMarginBottom: z.number().optional(),
    itemMarginTop: z.number().optional(),
    itemStyle: z.any().optional(),
    itemWidth: z.number().optional(),
    labelFormat: z.string().optional(),
    labelFormatter: z.any().optional(),
    layout: z.any().optional(),
    margin: z.number().optional(),
    maxHeight: z.number().optional(),
    maxWidth: z.any().optional(),
    navigation: legendNavigationOptionsSchema.optional(),
    padding: z.number().optional(),
    reversed: z.boolean().optional(),
    rtl: z.boolean().optional(),
    shadow: z.any().optional(),
    squareSymbol: z.boolean().optional(),
    symbolHeight: z.number().optional(),
    symbolPadding: z.number().optional(),
    symbolRadius: z.number().optional(),
    symbolWidth: z.number().optional(),
    title: legendTitleOptionsSchema.optional(),
    useHTML: z.boolean().optional(),
    valueDecimals: z.number().optional(),
    valueSuffix: z.string().optional(),
    verticalAlign: z.any().optional(),
    width: z.any().optional(),
    x: z.number().optional(),
    y: z.number().optional()
});
// ...
export const subtitleOptionsSchema = z.object({
    align: z.any().optional(),
    floating: z.boolean().optional(),
    style: z.any().optional(),
    text: z.string().optional(),
    useHTML: z.boolean().optional(),
    verticalAlign: z.any().optional(),
    x: z.number().optional(),
    y: z.number().optional()
});

export const timeOptionsSchema = z.object({
    Date: z.any().optional(),
    timezone: z.string().optional(),
    timezoneOffset: z.number().optional()
});

export const titleOptionsSchema = z.object({
    align: z.any().optional(),
    floating: z.boolean().optional(),
    margin: z.number().optional(),
    minScale: z.number().optional(),
    style: z.any().optional(),
    text: z.string().optional(),
    useHTML: z.boolean().optional(),
    verticalAlign: z.any().optional(),
    x: z.number().optional(),
    y: z.number().optional()
});
// ...
export const tooltipOptionsSchema = z.object({
    animation: z.any().optional(),
    backgroundColor: z.any().optional(),
    borderColor: z.any().optional(),
    borderRadius: z.number().optional(),
    borderWidth: z.number().optional(),
    changeDecimals: z.number().optional(),
    className: z.string().optional(),
    clusterFormat: z.string().optional(),
    dateTimeLabelFormats: z.any().optional(),
    distance: z.number().optional(),
    enabled: z.boolean().optional(),
    fixed: z.boolean().optional(),
    followPointer: z.boolean().optional(),
    followTouchMove: z.boolean().optional(),
    footerFormat: z.string().optional(),
    format: z.string().optional(),
    formatter: z.any().optional(),
    headerFormat: z.string().optional(),
    headerShape: z.any().optional(),
    hideDelay: z.number().optional(),
    nullFormat: z.string().optional(),
    nullFormatter: z.any().optional(),
    outside: z.any().optional(),
    padding: z.number().optional(),
    pointFormat: z.string().optional(),
    pointFormatter: z.any().optional(),
    position: tooltipPositionOptionsSchema.optional(),
    positioner: z.any().optional(),
    shadow: z.any().optional(),
    shape: z.any().optional(),
    shared: z.boolean().optional(),
    snap: z.number().optional(),
    split: z.boolean().optional(),
    stickOnContact: z.boolean().optional(),
    style: z.any().optional(),
    useHTML: z.boolean().optional(),
    valueDecimals: z.any().optional(),
    valuePrefix: z.string().optional(),
    valueSuffix: z.string().optional(),
    xDateFormat: z.any().optional()
});
// ...
export const optionsSchema: z.ZodSchema<Options> = z.lazy(() => z.object({
    accessibility: accessibilityOptionsSchema.optional(),
    annotations: z.any().optional(),
    boost: boostOptionsSchema.optional(),
    caption: captionOptionsSchema.optional(),
    chart: chartOptionsSchema.optional(),
    colorAxis: z.any().optional(),
    colors: z.any().optional(),
    connectors: connectorsOptionsSchema.optional(),
    credits: creditsOptionsSchema.optional(),
    data: dataOptionsSchema.optional(),
    defs: z.any().optional(),
    drilldown: drilldownOptionsSchema.optional(),
    exporting: exportingOptionsSchema.optional(),
    global: globalOptionsSchema.optional(),
    lang: langOptionsSchema.optional(),
    legend: legendOptionsSchema.optional(),
    loading: loadingOptionsSchema.optional(),
    mapNavigation: mapNavigationOptionsSchema.optional(),
    mapView: mapViewOptionsSchema.optional(),
    navigation: navigationOptionsSchema.optional(),
    navigator: navigatorOptionsSchema.optional(),
    noData: noDataOptionsSchema.optional(),
    pane: z.any().optional(),
    plotOptions: plotOptionsSchema.optional(),
    rangeSelector: rangeSelectorOptionsSchema.optional(),
    responsive: responsiveOptionsSchema.optional(),
    scrollbar: scrollbarOptionsSchema.optional(),
    series: z.any().optional(),
    sonification: sonificationOptionsSchema.optional(),
    stockTools: stockToolsOptionsSchema.optional(),
    subtitle: subtitleOptionsSchema.optional(),
    time: timeOptionsSchema.optional(),
    title: titleOptionsSchema.optional(),
    tooltip: tooltipOptionsSchema.optional(),
    xAxis: z.any().optional(),
    yAxis: z.any().optional(),
    zAxis: z.any().optional()
}));